Move ApiMarkPoint and ApiUserMarkContainer to separate files.
