+++ map/user_mark_container.cpp
@@ -267,27 +267,6 @@ UserMark * SearchUserMarkContainer::AllocateUserMark(const m2::PointD & ptOrg)
   return new SearchMarkPoint(ptOrg, this);
 }
 
-ApiUserMarkContainer::ApiUserMarkContainer(double layerDepth, Framework & framework)
-  : UserMarkContainer(layerDepth, framework)
-{
-}
-
-string ApiUserMarkContainer::GetTypeName() const
-{
-  return "api-result";
-}
-
-string ApiUserMarkContainer::GetActiveTypeName() const
-{
-  return "search-result-active";
-}
-
-UserMark * ApiUserMarkContainer::AllocateUserMark(const m2::PointD & ptOrg)
-{
-  return new ApiMarkPoint(ptOrg, this);
-}
-
-
 DebugUserMarkContainer::DebugUserMarkContainer(double layerDepth, Framework & framework)
   : UserMarkContainer(layerDepth, framework)
 {
